Separated kernel logs into its own component
